refactor: Moved Word components
