+++ src/components/vocabulary/RulesModal.tsx
