+++ src/components/vocabulary/WordInfoModal.tsx
