+++ src/components/vocabulary/WordItem.tsx
