+++ src/components/vocabulary/WordSkill.tsx
