+++ src/components/vocabulary/WordVariant.tsx
